Bounce
Beginning stages to the bounce mechanic. Ants now bounce off walls in a semi natural way. It is not perfect and sometimes the ants get stuck in the corners. Right now I just have the different cases hardcoded. In the future I will want to do an actual calculation to reflect the ant. But that will require a little overhaul of the wall mechanic.